refactor: Status 의 생성자 파라미터 변경
- Status 외부에서 직접 계산 한 뒤에 결과 값으로 생성하도록 함
