+++ src/main/java/chess/domain/state/Running.java
@@ -1,7 +1,8 @@
 package chess.domain.state;
 
 import chess.domain.board.Board;
-import chess.domain.piece.notation.Color;
+import chess.domain.board.BoardCalculator;
+import chess.domain.piece.Color;
 import chess.domain.position.Position;
 
 public class Running extends State {
@@ -25,20 +26,20 @@ public class Running extends State {
 
     @Override
     public Status status() {
-        return new Status(Color.EMPTY, board.getValue());
+        final var boardCalculator = new BoardCalculator(board.getValue());
+        return new Status(Result.from(Color.EMPTY), boardCalculator.sumScore(Color.WHITE), boardCalculator.sumScore(Color.BLACK));
     }
 
     @Override
     public State move(final Position from, final Position to) {
         checkPosition(currentColor, from, to);
         board.move(from, to);
-
         final var nextColor = currentColor.next();
 
         if (board.hasKing(nextColor)) {
             return new Running(nextColor, board);
         }
-        return new Finish(board, currentColor);
+        return new Finish(currentColor, board);
     }
 
     private void checkPosition(final Color color, final Position from, final Position to) {
