+++ src/test/java/chess/domain/state/RunningTest.java
@@ -3,7 +3,7 @@ package chess.domain.state;
 import chess.domain.board.Board;
 import chess.domain.piece.Pawn;
 import chess.domain.piece.Piece;
-import chess.domain.piece.notation.Color;
+import chess.domain.piece.Color;
 import chess.domain.position.Position;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
